+++ src/lang/all.ts
@@ -1,5 +1,7 @@
 import { all as baseGrammars } from "@wooorm/starry-night";
 import sourceJS from "./source.js.js";
+import sourceTS from "./source.ts.js";
+import sourceTSX from "./source.tsx.js";
 
 /**
  * List of all grammars.
@@ -9,7 +11,16 @@ import sourceJS from "./source.js.js";
  *
  * @see {@link https://github.com/wooorm/starry-night#all}
  */
-const grammars = baseGrammars.filter((g) => g.scopeName !== "source.js");
-grammars.push(sourceJS);
+const grammars = baseGrammars.filter(({ scopeName }) => {
+  switch (scopeName) {
+    case "source.js":
+    case "source.ts":
+    case "source.tsx":
+      return false;
+  }
+
+  return true;
+});
+grammars.push(sourceJS, sourceTS, sourceTSX);
 
 export default grammars;