+++ src/lang/common.ts
@@ -1,5 +1,6 @@
 import { common as baseGrammars } from "@wooorm/starry-night";
 import sourceJS from "./source.js.js";
+import sourceTS from "./source.ts.js";
 
 /**
  * List of common grammars.
@@ -9,7 +10,15 @@ import sourceJS from "./source.js.js";
  *
  * @see {@link https://github.com/wooorm/starry-night#common}
  */
-const grammars = baseGrammars.filter((g) => g.scopeName !== "source.js");
-grammars.push(sourceJS);
+const grammars = baseGrammars.filter(({ scopeName }) => {
+  switch (scopeName) {
+    case "source.js":
+    case "source.ts":
+      return false;
+  }
+
+  return true;
+});
+grammars.push(sourceJS, sourceTS);
 
 export default grammars;